requirements.txt
amqp==5.2.0
appnope==0.1.3
arc4==0.4.0
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.8.2
camelot-py==0.11.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configparser==7.1.0
cryptography==43.0.3
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
fuzzywuzzy==0.18.0
ghostscript==0.7
greenlet==3.1.1
h11==0.16.0
idna==3.10
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kombu==5.4.2
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
Naked==0.1.32
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.3
pdfminer.six==20240706
pexpect==4.8.0
pip-autoremove==0.10.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.16.1
pypdf==5.0.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
selenium==4.31.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.36
stack-data==0.6.3
tabula-py==2.10.0
tabulate==0.9.0
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.9
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xlrd==2.0.1
zipp==3.20.2
